+++ packages/components-js/tests/vrt/puppeteer/specs/banner.vrt.ts
@@ -37,13 +37,13 @@ it('should have no visual regression for :hover + :focus-visible', async () => {
 
       await setContentWithDesignSystem(page, getThemedBodyMarkup(getElementsMarkup), { injectIntoHead: head });
 
-      await forceHoverState(page, '.hover > p-banner span a');
+      await forceHoverState(page, '.hover p-banner span a');
       // TODO: support for 3rd level of shadow DOM is missing
-      // await forceHoveredState(page, '.hover > p-banner >>> p-inline-notification >>> p-button-pure >>> button');
-      await forceFocusState(page, '.focus > p-banner span a');
-      // await forceFocusedState(page, '.focus > p-banner >>> p-inline-notification >>> p-button-pure >>> button');
-      await forceFocusHoverState(page, '.focus-hover > p-banner span a');
-      // await forceFocusedHoveredState(page, '.focus-hover > p-banner >>> p-inline-notification >>> p-button-pure >>> button');
+      // await forceHoveredState(page, '.hover p-banner >>> p-inline-notification >>> p-button-pure >>> button');
+      await forceFocusState(page, '.focus p-banner span a');
+      // await forceFocusedState(page, '.focus p-banner >>> p-inline-notification >>> p-button-pure >>> button');
+      await forceFocusHoverState(page, '.focus-hover p-banner span a');
+      // await forceFocusedHoveredState(page, '.focus-hover p-banner >>> p-inline-notification >>> p-button-pure >>> button');
     })
   ).toBeFalsy();
 });