+++ packages/storefront/src/pages/patterns/apps/example-carousel.vue
@@ -115,7 +115,7 @@
 </script>
 
 <style lang="scss" scoped>
-  @import '~@porsche-design-system/components-js/utilities/scss';
+  @import '~@porsche-design-system/components-js/styles/scss';
 
   main {
     @include pds-grid;